move the knight when the dices number is 7

+++ CatanGame/CatanGame/Game/Board/Structures/IStructure.h
@@ -12,7 +12,7 @@ public:
 		m_player(player), 
 		m_resources(std::move(resources)){}
 	const PlayerType& get_player() const { return m_player; }
-	virtual std::unordered_map<ResourceType, uint8_t> get_resources(const uint8_t dice_nubmer) = 0;
+	virtual std::unordered_map<ResourceType, uint8_t> get_resources(const uint32_t dice_nubmer) = 0;
 
 protected:
 	PlayerType m_player;
